Commit 05/11/2022

+++ Onlineshop/Program.cs
@@ -1,4 +1,5 @@
-﻿using Onlineshop.Model;
+﻿using Lieferservice;
+using Onlineshop.Model;
 using System;
 
 namespace Onlineshop
@@ -7,7 +8,16 @@ namespace Onlineshop
     {
         static void Main(string[] args)
         {
-            using (OnlineshopContext db = new OnlineshopContext())
+            if (args.Length < 1)
+            {
+                Console.Error.WriteLine("Missing args.");
+                Console.Error.WriteLine("Usage: dotnet run -- (sqlserver|oracle|sqlite)");
+                return;
+            }
+            var options = MultiDbContext.GetConnectionInteractive(dbms: args[0].ToLower(), database: "Onlineshop");
+            if (options is null) { return; }
+
+            using (OnlineshopContext db = new OnlineshopContext(options))
             {
                 db.Database.EnsureDeleted();
                 if (db.Database.EnsureCreated()) { db.Seed(); }
